requirements.txt
numpy
absl-py
tensorflow-gpu==1.6.0
scipy
pyflann
